fixed typing issues
